Finished removing app-assigned "local" event sequence and only using the "global" (database-assigned) event sequence for ordering (now just called "event_sequence)
Since event_sequence is a BIGSERIAL, we no longer need to have a composite ID. However, we still index on aggregate_root_id as that's often how we'll retrieve events, e.g., to reconstitute an Aggregate

ALL TESTS PASS

+++ src/main/java/dev/ted/jitterticket/eventsourced/domain/concert/ConcertEvent.java
@@ -7,7 +7,7 @@ public abstract sealed class ConcertEvent extends Event
 
     private final ConcertId concertId;
 
-    protected ConcertEvent(ConcertId concertId, Integer eventSequence) {
+    protected ConcertEvent(ConcertId concertId, Long eventSequence) {
         super(eventSequence);
         this.concertId = concertId;
     }
